+++ CHANGELOG.md
@@ -595,3 +595,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix calculateModifiedCardDamage (kiooeht)
 
 #### dev ####
+* Allow custom card tooltips before normal tooltips (kiooeht)